# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Consultation {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  interest: string;
  /** @nullable */
  message?: string | null;
  status: string;
  createdAt: Date;
  updatedAt: Date;
}