# lib/api-zod/src/generated/types/consultationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultationInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  interest: string;
  message?: string;
}