# lib/api-zod/src/generated/types/inventoryItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  lowStockThreshold: number;
  notes?: string;
}