# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  /** @minimum 0 */
  stock?: number;
  unit?: string;
  /** @minimum 0 */
  lowStockThreshold?: number;
  notes?: string;
}